fix: 코드리뷰 CRITICAL/HIGH 이슈 10건 수정 및 결제 롤백 버그 해결

See merge request s14-fintech-finance-sub1/S14P21C203!42

+++ BE/src/main/java/com/ssafy/tax7i/taxcalendar/service/TaxCalendarService.java
@@ -39,12 +39,12 @@ public class TaxCalendarService {
             result.add(toResponse(d, deadline, today));
         }
 
-        result.sort(Comparator.comparingLong(TaxDeadlineResponse::dDay));
+        result.sort(Comparator.comparingInt(TaxDeadlineResponse::dDay));
         return result;
     }
 
     private TaxDeadlineResponse toResponse(TaxDeadline d, LocalDate deadline, LocalDate today) {
-        long dDay = ChronoUnit.DAYS.between(today, deadline);
+        int dDay = (int) ChronoUnit.DAYS.between(today, deadline);
         return new TaxDeadlineResponse(d.getName(), d.getDescription(), deadline.toString(), dDay);
     }
 }
